+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.0.0-beta3] - In progress
+
+### Added
+
+- New method `DefaultHeadersTrait::addDefaultHeaders`
+
+### Fixed
+
+- Redirection keep old headers like `Content-Length`
+
 ## [2.0.0-beta2] - 2022-01-13
 
 ### Added